requirements.txt
speechrecognition>=3.10.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
google-generativeai>=0.8.3
pyttsx3>=2.90